Pandas/pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pandas-mcp"
version = "0.1.0"
description = "Pandas MCP server implementation using Model Context Protocol with comprehensive data analysis capabilities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.1.0",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.21.0",
    "pydantic>=1.10.0",
    "pytest>=7.2.0",
    "requests>=2.28.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.0",
    "pyarrow>=15.0.0",
    "tables>=3.9.0",
    "sqlalchemy>=2.0.0",
    "psutil>=5.9.0",
    "scikit-learn>=1.3.0"
]

[project.optional-dependencies]
test = ["pytest>=7.0.0", "httpx"]

[project.scripts]
pandas-mcp = "pandasmcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["pandas*"]

[tool.setuptools.package-dir]
"pandas" = "src/pandas"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

[tool.isort]
profile = "black"
line_length = 88
